requirements.txt
python-pptx
requests
BeautifulSoup
markdownify
openai
dotenv
moviepy